+++ packages/video_player/video_player/test/video_player_test.dart
@@ -84,6 +84,11 @@ class FakeController extends ValueNotifier<VideoPlayerValue>
   Future<void> setClosedCaptionFile(
     Future<ClosedCaptionFile>? closedCaptionFile,
   ) async {}
+
+  @override
+  Future<void> updateNotificationMetadata(
+    NotificationMetadata notificationMetadata,
+  ) async {}
 }
 
 Future<ClosedCaptionFile> _loadClosedCaption() async =>
@@ -1625,5 +1630,27 @@ class FakeVideoPlayerPlatform extends VideoPlayerPlatform {
     }
     calls.add('setWebOptions');
     webOptions[playerId] = options;
+  }
+
+  @override
+  bool isBackgroundPlaybackSupportAvailable() {
+    return true;
+  }
+
+  @override
+  Future<void> setBackgroundPlayback(
+    int playerId, {
+    required bool enableBackground,
+    NotificationMetadata? notificationMetadata,
+  }) async {
+    calls.add('setBackgroundPlayback');
+  }
+
+  @override
+  Future<void> updateNotificationMetadata(
+    int playerId,
+    NotificationMetadata notificationMetadata,
+  ) async {
+    calls.add('updateNotificationMetadata');
   }
 }