+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 6.7.0
 
+* Adds background playback with system media notification support.
 * Updates minimum supported SDK version to Flutter 3.32/Dart 3.8.
 
 ## 6.6.0